#3583 comment: added test_clean.py
